Adds clean method to add game form (#943)
* Added clean name method to add game form.

* Bumped up version number.

* Added tests.

* Splits test files per class.

* Moved playable games from setter to init

* Merge branch 'development' into clean-add-game-form

# Conflicts:
#	aimmo/tests/tests_views.py
#	version.txt

* Moves helper to top of function
